+++ libs/components/pages/projects/pages/src/modules/action-hub/fixtures/action-hub-content-fixture.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'sky-action-hub-content-fixture',
+  templateUrl: 'action-hub-content-fixture.component.html'
+})
+export class ActionHubContentFixtureComponent {
+  public title = 'Page title';
+  public label = 'world';
+  public value = 'hello';
+}