+++ libs/components/pages/projects/pages/src/public-api.ts
@@ -1,0 +1,12 @@
+export * from './modules/action-hub/action-hub.module';
+export * from './modules/action-hub/types/action-hub-needs-attention';
+export * from './modules/action-hub/types/page-link';
+export * from './modules/action-hub/types/recent-link';
+export * from './modules/page-header/page-header.module';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of the public API.
+export {SkyActionHubComponent as λ1} from './modules/action-hub/action-hub.component';
+export {SkyActionHubButtonsComponent as λ2} from './modules/action-hub/action-hub-buttons.component';
+export {SkyActionHubContentComponent as λ3} from './modules/action-hub/action-hub-content.component';
+export {SkyPageHeaderComponent as λ4} from './modules/page-header/page-header.component';