+++ libs/components/pages/src/app/public/plugin-resources/pages-resources-provider.ts
@@ -1,5 +1,0 @@
-import { SkyAppLocaleInfo, SkyLibResourcesProvider } from '@skyux/i18n';
-
-export class SkyPagesResourcesProvider implements SkyLibResourcesProvider {
-  public getString: (localeInfo: SkyAppLocaleInfo, name: string) => string;
-}