+++ libs/components/pages/src/app/public/public_api.ts
@@ -1,5 +1,0 @@
-export * from './modules/action-hub/action-hub.module';
-export * from './modules/action-hub/types/action-hub-needs-attention';
-export * from './modules/action-hub/types/page-link';
-export * from './modules/action-hub/types/recent-link';
-export * from './modules/page-header/page-header.module';